react router with createBrowserRouter
